Split PR catalog validator into modules
